requirements.txt
Django~=5.0.1
django-ckeditor~=6.7.0